refactor(frontend): feature-based restructure, i18n fixes, responsive lobby
- Reorganise src/ into feature folders: game/modes/{solo,local,online}, features/{auth,settings,profile}, components/overlays
- Split large files: OnlineMode (551→ multi-fichiers), ProfilePage (825→ hook+composants), GlobalOverlays (358→ 3 overlays), useWebSocket (358→ +useChatWebSocket), useAuth (298→ authApi+useLogin/Register/Logout), ResultsPage (→ +AnimatedBar)
- Move pages/ domain files to their feature folders (settings, game, profile)
- Fix DEFAULT_AVATAR: use Vite ?url import via utils/defaultAvatar.ts across all components
- i18n: add missing keys in EN/FR/ES (Private Message, Aucune bio, Send invite, Theme, Time, Your character, bio login errors, Failed to fetch channels)
- Responsive lobby: replace absolute Back button with header row; switch min-h/overflow-auto to h-[100dvh]/overflow-hidden with clamp-based spacing so pages never scroll
- Delete stale test assets (test.png, test.svg, vite.svg, Zone.Identifier files)
