--- srcs/frontend/public/Icône dynamique de football en mouvement (1).svg:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://www.iloveimg.com/
-HostUrl=https://api1.iloveimg.com/v1/download/yxw23my363lpt1ld4xA614Atvh13c8vnbj3tt4yy3rzh7sxsm1vqgh215ynA8z75g2ph4wmyt815mr1xh57A25ft42f3zrfqyrrk1h545y8k33k68A0vnyyzA30lf1jA9756x3ghbg9hhgfr2gsn278fv7A852zcfpt4m5xApA1rvfh7qdz1
