+++ srcs/frontend/public/locales/en/translation.json
@@ -119,6 +119,7 @@
     "Failed to delete room": "Failed to delete room",
     "Chat": "Chat",
     "Messages directs": "Messages directs",
+    "Messages": "Messages",
     "Refresh": "Refresh",
     "Close": "Close",
     "Room name": "Room name",
@@ -311,5 +312,34 @@
     "Restart": "Restart",
     "Abandon": "Abandon",
     "abandon.confirm_title": "Abandon match?",
-    "abandon.confirm_desc": "You will forfeit and lose LP."
+    "abandon.confirm_desc": "You will forfeit.",
+
+    "Direct Messages": "Direct Messages",
+    "No direct messages yet.": "No direct messages yet.",
+    "Select a conversation": "Select a conversation",
+    "Join the channel to chat": "Join the channel to chat",
+    "Delete": "Delete",
+    "Today": "Today",
+    "Yesterday": "Yesterday",
+    "Offline": "Offline",
+    "goals": "goals",
+    "just now": "just now",
+    "min ago": "min ago",
+    "ago": "ago",
+    "s": "s",
+    "Cannot send message to blocked user": "Cannot send message to blocked user",
+    "Unable to leave room": "Unable to leave room",
+    "Failed to start conversation": "Failed to start conversation",
+    "Failed to create group": "Failed to create group",
+    "Failed to invite member": "Failed to invite member",
+    "Failed to fetch messages": "Failed to fetch messages",
+    "Private Message": "Private Message",
+    "Aucune bio": "No bio",
+    "Failed to fetch channels": "Failed to fetch channels",
+    "Send invite": "Send invite",
+    "Theme": "Theme",
+    "Time": "Time",
+    "Your character": "Your character",
+    "You must be logged in to edit your bio.": "You must be logged in to edit your bio.",
+    "You must be logged in to save your bio.": "You must be logged in to save your bio."
 }
