+++ srcs/frontend/public/locales/es/translation.json
@@ -147,6 +147,9 @@
     "Write a message...": "Escribir un mensaje...",
     "Send": "Enviar",
     "No channel selected": "No se ha seleccionado un canal",
+    "Messages": "Mensajes",
+    "Failed to fetch messages": "No se pudieron obtener los mensajes",
+    "Failed to fetch channels": "No se pudieron obtener los canales",
     "Back home": "Volver a inicio",
 
     "GOAAAAL !": "¡GOAAAAL!",
@@ -300,5 +303,45 @@
     "Restart": "Reiniciar",
     "Abandon": "Abandonar",
     "abandon.confirm_title": "¿Abandonar la partida?",
-    "abandon.confirm_desc": "Perderás la partida y los LP."
+    "abandon.confirm_desc": "Perderás la partida.",
+
+    "Direct Messages": "Mensajes directos",
+    "No direct messages yet.": "Aún no hay mensajes directos.",
+    "Select a conversation": "Seleccionar una conversación",
+    "Join the channel to chat": "Únete al canal para chatear",
+    "Delete": "Eliminar",
+    "Today": "Hoy",
+    "Yesterday": "Ayer",
+    "Offline": "Desconectado",
+    "goals": "goles",
+    "just now": "ahora mismo",
+    "min ago": "min",
+    "ago": "",
+    "s": "s",
+    "Cannot send message to blocked user": "No se puede enviar un mensaje a un usuario bloqueado",
+    "Unable to leave room": "No se puede salir del canal",
+    "Failed to start conversation": "No se pudo iniciar la conversación",
+    "Failed to create group": "No se pudo crear el grupo",
+    "Failed to invite member": "No se pudo invitar al miembro",
+    "Failed to fetch messages": "No se pudieron obtener los mensajes",
+    "Public channels": "Canales públicos",
+    "Groups": "Grupos",
+    "Group name": "Nombre del grupo",
+    "Invite friends": "Invitar amigos",
+    "No public channels yet.": "Aún no hay canales públicos.",
+    "No groups yet.": "Aún no hay grupos.",
+    "Invite a friend": "Invitar a un amigo",
+    "All friends are already in this group.": "Todos los amigos ya están en este grupo.",
+    "m": "m",
+    "Channel name": "Nombre del canal",
+    "Create group": "Crear grupo",
+    "Rooms": "Salas",
+    "Private Message": "Mensaje privado",
+    "Aucune bio": "Sin biografía",
+    "Send invite": "Enviar invitación",
+    "Theme": "Tema",
+    "Time": "Tiempo",
+    "Your character": "Tu personaje",
+    "You must be logged in to edit your bio.": "Debes iniciar sesión para editar tu biografía.",
+    "You must be logged in to save your bio.": "Debes iniciar sesión para guardar tu biografía."
 }
