+++ srcs/frontend/public/locales/fr/translation.json
@@ -120,6 +120,7 @@
     "Failed to delete room": "Échec de la suppression de la salle",
     "Chat": "Chat",
     "Messages directs": "Messages directs",
+    "Messages": "Messages",
     "Refresh": "Actualiser",
     "Close": "Fermer",
     "Room name": "Nom de la salle",
@@ -312,6 +313,35 @@
     "Restart": "Recommencer",
     "Abandon": "Abandonner",
     "abandon.confirm_title": "Abandonner la partie ?",
-    "abandon.confirm_desc": "Vous perdrez la partie et des LP."
+    "abandon.confirm_desc": "Vous perdrez la partie.",
+
+    "Direct Messages": "Messages directs",
+    "No direct messages yet.": "Aucun message direct pour l'instant.",
+    "Select a conversation": "Sélectionner une conversation",
+    "Join the channel to chat": "Rejoindre le canal pour discuter",
+    "Delete": "Supprimer",
+    "Today": "Aujourd'hui",
+    "Yesterday": "Hier",
+    "Offline": "Hors ligne",
+    "goals": "buts",
+    "just now": "à l'instant",
+    "min ago": "min",
+    "ago": "",
+    "s": "s",
+    "Cannot send message to blocked user": "Impossible d'envoyer un message à un utilisateur bloqué",
+    "Unable to leave room": "Impossible de quitter le canal",
+    "Failed to start conversation": "Échec du démarrage de la conversation",
+    "Failed to create group": "Échec de la création du groupe",
+    "Failed to invite member": "Échec de l'invitation du membre",
+    "Failed to fetch messages": "Échec de la récupération des messages",
+    "Private Message": "Message privé",
+    "Aucune bio": "Aucune bio",
+    "Failed to fetch channels": "Échec du chargement des salons",
+    "Send invite": "Envoyer l'invitation",
+    "Theme": "Thème",
+    "Time": "Temps",
+    "Your character": "Votre personnage",
+    "You must be logged in to edit your bio.": "Vous devez être connecté pour modifier votre bio.",
+    "You must be logged in to save your bio.": "Vous devez être connecté pour sauvegarder votre bio."
 }
