+++ srcs/frontend/src/features/game/modes/online/constants.ts
@@ -1,0 +1,11 @@
+export const PLAYER1_COLOR = '#3b82f6'
+export const PLAYER2_COLOR = '#ef4444'
+export const FORFEIT_TIMEOUT_S = 20
+
+export const ARCADE_BTN = '0px 4px rgb(255,255,255), 0px -4px rgb(255,255,255), 4px 0px rgb(255,255,255), -4px 0px rgb(255,255,255), 0px 4px rgba(0,0,0,0.22), 4px 4px rgba(0,0,0,0.22), -4px 4px rgba(0,0,0,0.22), inset 0px 4px rgba(255,255,255,0.21)'
+
+export const ARCADE_BASE =
+  'hb-tap font-arcade border-0 shadow-[0px_4px_rgb(255,255,255),0px_-4px_rgb(255,255,255),4px_0px_rgb(255,255,255),-4px_0px_rgb(255,255,255),0px_4px_rgba(0,0,0,0.22),4px_4px_rgba(0,0,0,0.22),-4px_4px_rgba(0,0,0,0.22),inset_0px_4px_rgba(255,255,255,0.21)] cursor-pointer no-underline inline-flex items-center justify-center transition-transform duration-100 active:translate-y-0.5'
+
+export const ONLINE_SCORE = 5
+export const ONLINE_DURATION = 120
