+++ srcs/frontend/src/utils/defaultAvatar.ts
@@ -1,0 +1,3 @@
+import defaultAvatarUrl from '/assets/default-profile.jpg?url'
+
+export const DEFAULT_AVATAR = defaultAvatarUrl
